+++ src/vscode-ui-extension/tests/windowsStore.test.ts
@@ -10,7 +10,11 @@ import { buildPfx } from "../src/cert/pfx";
 import { type DevCert, type DevKey } from "../src/cert/types";
 
 
-vi.mock("../src/platform/processUtil", () => ({
+// Mock runProcess at the shared internal path — WindowsCertificateStore
+// lives in `@devcontainer-dev-certs/shared/src/platform/windowsStore` after
+// the platform-layer move and imports `runProcess` from the sibling
+// `./processUtil`, so the mock must target that file.
+vi.mock("@devcontainer-dev-certs/shared/src/platform/processUtil", () => ({
   runProcess: vi.fn(),
 }));
 
@@ -19,7 +23,7 @@ import {
   type PsCandidate,
   type PsSkipped,
 } from "../src/platform/windowsStore";
-import { runProcess } from "../src/platform/processUtil";
+import { runProcess } from "@devcontainer-dev-certs/shared/src/platform/processUtil";
 
 const mockedRunProcess = vi.mocked(runProcess);
